This version supports parameter changes!!!!
All user-defined parameters in the data model are displayed with their
default values. The user can change any of them and run to see the
results. However, because the original data model is then re-read, the
parameter values will change back to defaults after each run. This needs
to be addressed ... later.