Attempt to use class to show and hide the end and step fields.
For some reason, changing the class on the fly appears to cause the
browser to break the line. An earlier version used a span element to
containt the optional end and step fields, but that also suffered from
the line breaking. This version changed the class in the input elements
themselves to hopefully get around the line breaking problem. But since
it didn't work, and the span element is a cleaner solution, the earlier
version will be committed after this one (effectively a roll back to an
earlier version that wasn't committed).