This version uses a span element to hide and show the end and step
As mentioned in the previous commit, changing the class of the input
fields seemed to insert a line break. This version reverts back to
wrapping the end and step fields (as well as text) in a span which is
hidden or visible. This results in a single line break instead of one
for each of the input fields. This is good enough for now.