Completed the parameter sweep looping ... need to handle subdirectories
This version sweeps through all parameters and the seed. It runs MCell
for each unique case. However, the results are written over the same
files for each case, so there's only one case left in the end.

The next step is to create the subdirectories and place the MDL in the
correct subdirectories and run MCell from each of those locations. That
will eventually be replaced with a job list, but the best way to be sure
that the job list is correct ... is to actually run it.