Added a run_parameters.json output file.
Each run will generate a JSON file containing a top level dictionary
with keys that match the parameters being generated. The values that
each parameter takes is included in a list which represents the value
for that key.

Note that this might not be the best representation because the keys in
a dictionary are not ordered (even though they appear to be in the JSON
file itself). The order is somewhat important because the subdirectory
structure containing the data depends on the order (first parameter is
at the top level, second parameter at the second, etc).

More discussion with Tom might be helpful.