Import of sweep parameters works, but needs to be loaded twice.
Some of the sweep information seems to be retained in the form. This is
not surprising since that was the only place it had been stored in the
previous version that didn't look for sweep information in the data
model. A little investigation might reveal the problem, but for now,
it's helpful to push the "Refresh" button after loading a data model to
get the proper number of runs. Otherwise it seems to remember sweep
settings from the previous form settings.