Updated to run properly using "run_data_model_mcell.py" from web
This version properly runs a data model by using the same run code that
is used from CellBlender (run_data_model_mcell.py). This is a interim
step in using that same code from within Galaxy.

Note that there were some relative/absolute path issues that were difficult
to resolve in a web environment. A simple "fake_mcell" was generated
that reported its parameters to help with debugging. That code is
included below.

This commit also included a number of JSON data models used in testing
the sweeping behavior. Note that not all of those data models can be
properly represented in this simple web interface because it limits
sweeps to start,end,step rather than a list of start,end,step values as
allowed in CellBlender. Those features will be supported in the Galaxy
version.

====== fake_mcell.c ======

int main ( int argc, char *argv[] ) {

  // ##### Start by reading the command line parameters which includes
  // the data model file name
  int i;
  char cwd[1024];
  getcwd(cwd, sizeof(cwd));
  printf ( "\nFake MCell running from %s\n  # args = %d", cwd, argc );
  for (i=0; i<argc; i++) {
    printf ( "\n    arg %d = %s", i, argv[i] );
  }
  printf ( "\n\n" );

}