Attempt to rebuild the parameter data model from the form data.
This code attempts to build a data model from the data on the form
directly rather than re-read the data model and update it.

This is being abandoned because the form data does not normally contain
the text items like units and description. These could be added to the
form as hidden items, but this seems like more work than needed. To keep
the code as simple as possible, the data model will be re-read and then
updated from the current form data.