This version appears to be working properly.
This version is quite a bit simpler than earlier versions because it
leverages the code in "run_data_model_mcell.py".

The one thing lost by delegating the "smarts" to run_data_model_mcell.py
is the ability to give instant feedback regarding the number of runs.
This could be handled by:

  1. Adding a limit option to "run_data_model_mcell.py"
    - or -
  2. Adding a count runs option to "run_data_model_mcell.py"

The first option would be easiest, and it would return an error message
if the number of runs exceeded the limit. That error message could be
displayed at the conclusion of the run.

The second option would be a little more difficult, but it would give a
nicer user interface. The only down side is that it would have to be
evaluated every time a change is made. The current application doesn't
have a mechanism to make a call on every parameter change. For this
reason, the first option will be implemented ... first.