Updated run_mcell_dm to show the runs generated by each parameter.
When a parameter generates more than one sample point, this version
shows the number of sample points generated beside the check box. This
version also updates the run button based on whether the total runs
exceeds the run limit. When the limit is exceeeded, the run button is
disabled.

Note that there is still a bit of "clunkiness" with this interface
because the JavaScript updating is attached to the "onchange" event of
the text fields. The "change" isn't detected until the focus leaves the
field. So the user must tab to a new field to see the effect. There was
an experimental version that attached the callback to the keypress
event, but there were some other problems with that version (more
exploration might help).

Using the tab key to complete fields is less natural than using the
"enter" key. However, the enter key will submit the form which isn't the
desired behavior either. The interface could benefit from disabling the
default enter-key submission.

Having said all of that, this version is actually pretty good as it is.